requirements.txt
seleniumbase>=4.15.0
fake-useragent>=1.4.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
newspaper3k==0.2.8
pyautogui==0.9.54
python-dotenv==1.0.0
tqdm==4.66.1
aiofiles==23.2.1
tenacity>=8.2.3
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.3